+++ changelog.md
@@ -3,6 +3,8 @@
 # Changelog #
 ## Changes in release 5.2.0 ##
 + Fixed issue where customer id was not added when using saved credit card for order API.
++ Added QR code option for bancontact payment on live mode
++ Added in3 payment method restriction for using only order API.
 
 ## Changes in release 5.1.0 ##
 + Added order refund statuses for methods that use Order API.